day2

+++ day.h
@@ -1,5 +1,5 @@
 #include <stddef.h>
 #define STR(x) #x
 #define XSTR(x) STR(x)
-int part1(size_t size, char *buffer);
-int part2(size_t size, char *buffer);
+long long part1(size_t size, char *buffer);
+long long part2(size_t size, char *buffer);
